requirements.txt
GitPython
watchdog
flask
PyYAML
